requirements.txt
streamlit>=1.28.0
anthropic>=0.40.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
